Completed implementation of logging for permissions.

+++ application/models/command_model.php
@@ -181,7 +181,7 @@ class Command_model extends CI_Model {
 		}
 
 		return array(
-			'trigger' => $this->_ci->Trigger_model->fetch($trigger_version->tid, $trigger_version->t_version),
+			'trigger' => $this->_ci->Trigger_model->fetch($trigger_version->tid, $trigger_version->version),
 			'permissions' => $this->fetch_permissions($rid, $tid),
 		);
 	}
@@ -214,7 +214,7 @@ class Command_model extends CI_Model {
 		$perm_list = array();
 		$filter = array(
 			'tid'                         => $trigger->tid,
-			't_version'                   => $trigger->t_version,
+			't_version'                   => $trigger->version,
 			'cmd_trigger_perm_map.status' => 1,
 		);
 		$query = $this->db->select('cmd_perm.*')
@@ -527,6 +527,10 @@ class Command_model extends CI_Model {
 
         $this->db->trans_begin();
 		try {
+			if(!$this->_ci->Trigger_model->delete($trigger->tid)) {
+				throw new Exception('Failed deleting the trigger');
+			}
+
 			if(!$this->_ci->Release_Trigger_Map_model->delete(array('rid' => $rid, 'tid' => $tid))) {
 				throw new Exception();
 			}
@@ -581,7 +585,7 @@ class Command_model extends CI_Model {
 				throw new Exception('failed to update rtm');
 			}
 
-			$mapping = $this->_ci->Trigger_Perm_Map_model->fetch(array('tid' => $tid, 't_version' => $current_trigger->t_version));
+			$mapping = $this->_ci->Trigger_Perm_Map_model->fetch(array('tid' => $tid, 't_version' => $current_trigger->version));
 			foreach($mapping as $map) {
 				$new_map = clone $map;
 				$new_map->t_version = $trigger->version;
@@ -592,7 +596,7 @@ class Command_model extends CI_Model {
 
 			$this->_log_event(Log_model::EVENT_RESTORE,
 			                  sprintf("Trigger %d from Release '%s' (rid: %d) reverted from version %d to %d",
-			                          $trigger->tid, $release->name, $release->rid, $current_trigger->t_version, $version));
+			                          $trigger->tid, $release->name, $release->rid, $current_trigger->version, $version));
 		}
 		catch (Exception $e) {
 			$this->db->trans_rollback();
@@ -625,7 +629,7 @@ class Command_model extends CI_Model {
 		if(empty($trigger_list)) {
 			return FALSE;
 		}
-		$trigger = $trigger_list[0];
+		$trigger = reset($trigger_list);
 
 		try {
 			$this->db->trans_begin();
@@ -645,6 +649,10 @@ class Command_model extends CI_Model {
 			if(!$this->_ci->Trigger_Perm_Map_model->insert($map)) {
 				throw new Exception();
 			}
+
+			$this->_log_event(Log_model::EVENT_INSERT,
+			                  sprintf("Permission '%s' added to Release '%s' (rid: %d), Trigger '%s' (tid: %d)",
+			                          $data['perm'], $release->name, $release->rid, $trigger->trigger, $trigger->tid));
 		}
 		catch (Exception $e) {
 			$this->db->trans_rollback();
@@ -678,7 +686,7 @@ class Command_model extends CI_Model {
 		if(empty($trigger_list)) {
 			return FALSE;
 		}
-		$trigger = $trigger_list[0];
+		$trigger = reset($trigger_list);
 
 		$mapping = $this->_ci->Trigger_Perm_Map_model->fetch(array('tid' => $trigger->tid, 't_version' => $trigger->version));
 		if(empty($mapping) || empty($mapping[$pid])) {
@@ -694,16 +702,23 @@ class Command_model extends CI_Model {
 		try {
 			$this->db->trans_begin();
 
-			$perm = $this->_ci->Perm_model->edit($perm->pid, $perm->version, $data);
-			if(empty($perm)) {
+			$updated_perm = $this->_ci->Perm_model->edit($perm->pid, $perm->version, $data);
+			if(empty($updated_perm)) {
 				throw new Exception();
 			}
 
 			$updated_map = clone $map;
-			$updated_map->p_version = $perm->version;
+			$updated_map->p_version = $updated_perm->version;
 			if(!$this->_ci->Trigger_Perm_Map_model->edit($updated_map, (array)$map)) {
 				throw new Exception();
 			}
+
+			$message = array();
+			foreach($data as $key => $value) {
+				$message[] = sprintf("%s: '%s' => '%s'", $key, $perm->{$key}, $value);
+			}
+			$this->_log_event(Log_model::EVENT_EDIT, sprintf("Permission %d for Release '%s' (rid: %d), Trigger '%s' (tid: %d): %s",
+			                                                 $perm->pid, $release->name, $release->rid, $trigger->trigger, $trigger->tid, implode(', ', $message)));
 		}
 		catch (Exception $e) {
 			$this->db->trans_rollback();
@@ -711,7 +726,7 @@ class Command_model extends CI_Model {
 		}
 
 		$this->db->trans_commit();
-		return $perm;
+		return $updated_perm;
 	}
 
 	/**
@@ -738,7 +753,7 @@ class Command_model extends CI_Model {
 			return FALSE;
 		}
 
-		$mapping = $this->_ci->Trigger_Perm_Map_model->fetch(array('tid' => $trigger->tid, 't_version' => $trigger->t_version));
+		$mapping = $this->_ci->Trigger_Perm_Map_model->fetch(array('tid' => $trigger->tid, 't_version' => $trigger->version));
 		if(empty($mapping[$pid])) {
 			return FALSE;
 		}
@@ -746,14 +761,24 @@ class Command_model extends CI_Model {
 
 		$perm = $this->_ci->Perm_model->fetch($pid, $map->p_version);
 
-        $this->db->trans_begin();
 		try {
-			if(!$this->_ci->Trigger_Perm_Map_model->delete((array)$map)) {
-				throw new Exception();
+			$this->db->trans_begin();
+
+			if(!$this->_ci->Perm_model->delete($perm->pid, $perm->version)) {
+				throw new Exception('Failed to delete the specified permission');
 			}
+
+			if(!$this->_ci->Trigger_Perm_Map_model->delete((array)$map)) {
+				throw new Exception('Filed to update TPM');
+			}
+
+			$this->_log_event(Log_model::EVENT_DELETE,
+			                  sprintf("Permission '%s' (pid: %d) deleted from Release '%s' (rid: %d), Trigger '%s' (tid: %d)",
+			                          $perm->perm, $perm->pid, $release->name, $release->rid, $trigger->trigger, $trigger->tid));
 		}
 		catch (Exception $e) {
 			$this->db->trans_rollback();
+			echo $e->getMessage();
 			return FALSE;
 		}
 
@@ -786,6 +811,11 @@ class Command_model extends CI_Model {
 			return FALSE;
 		}
 
+		$current_perm = $this->_ci->Perm_model->fetch($pid);
+		if(empty($current_perm)) {
+			return FALSE;
+		}
+
 		try {
 			$this->db->trans_begin();
 
@@ -796,12 +826,12 @@ class Command_model extends CI_Model {
 
 			$updated_map = array(
 				'tid'       => $trigger->tid,
-				't_version' => $trigger->t_version,
+				't_version' => $trigger->version,
 				'pid'       => $perm->pid,
 				'p_version' => $perm->version,
 			);
 
-			$current_map = $this->_ci->Trigger_Perm_Map_model->fetch(array('tid' => $trigger->tid, 't_version' => $trigger->t_version, 'pid' => $pid));
+			$current_map = $this->_ci->Trigger_Perm_Map_model->fetch(array('tid' => $trigger->tid, 't_version' => $trigger->version, 'pid' => $pid));
 			if(empty($current_map[$pid])) {
 				if(!$this->_ci->Trigger_Perm_Map_model->insert($updated_map)) {
 					throw new Exception('Failed updating empty TPM');
@@ -811,6 +841,9 @@ class Command_model extends CI_Model {
 				throw new Exception('Failed updating TPM');
 			}
 
+			$this->_log_event(Log_model::EVENT_RESTORE,
+			                  sprintf("Permission %d for Release '%s' (rid: %d), Trigger '%s' (tid: %d) restored from version %d to %d",
+			                          $perm->pid, $release->name, $release->rid, $trigger->trigger, $trigger->tid, $current_perm->version, $version));
 		}
 		catch(Exception $e) {
 			$this->db->trans_rollback();
@@ -999,7 +1032,7 @@ class Command_model extends CI_Model {
 		$mapping = $this->_ci->Release_Trigger_Map_model->fetch($rid);
 		foreach($mapping as $trigger) {
 			if($trigger->tid == $tid) {
-				return $trigger;
+				return $this->_ci->Trigger_model->fetch($trigger->tid, $trigger->t_version);
 			}
 		}
